fixed errors given by -Wall and -Wextra

+++ project/airport/runway.h
@@ -8,25 +8,25 @@ enum Runway_activity {idle, land, takeoff};
 
 class Runway {
 public:
-   Runway(int limit);
-   Error_code can_land(Plane current);
-   Error_code can_depart(Plane current);
-   Runway_activity activity(int time, Plane &moving);
-   void shut_down(int time) const;
+	Runway(int limit);
+	Error_code can_land(Plane current);
+	Error_code can_depart(Plane current);
+	Runway_activity activity(int time, Plane &moving);
+	void shut_down(int time) const;
 
 private:
-   Queue<Plane> landingQ;
-   Queue<Plane> takeoffQ;
-   int queue_limit;
-   int num_land_requests;        //  number of planes asking to land
-   int num_takeoff_requests;     //  number of planes asking to take off
-   int num_landings;             //  number of planes that have landed
-   int num_takeoffs;             //  number of planes that have taken off
-   int num_land_accepted;        //  number of planes queued to land
-   int num_takeoff_accepted;     //  number of planes queued to take off
-   int num_land_refused;         //  number of landing planes refused
-   int num_takeoff_refused;      //  number of departing planes refused
-   int land_wait;                //  total time of planes waiting to land
-   int takeoff_wait;             //  total time of planes waiting to take off
-   int idle_time;                //  total time runway is idle
+	Queue<Plane> landingQ;
+	Queue<Plane> takeoffQ;
+	size_t queue_limit;
+	size_t num_land_requests;        //  number of planes asking to land
+	size_t num_takeoff_requests;     //  number of planes asking to take off
+	size_t num_landings;             //  number of planes that have landed
+	size_t num_takeoffs;             //  number of planes that have taken off
+	size_t num_land_accepted;        //  number of planes queued to land
+	size_t num_takeoff_accepted;     //  number of planes queued to take off
+	size_t num_land_refused;         //  number of landing planes refused
+	size_t num_takeoff_refused;      //  number of departing planes refused
+	size_t land_wait;                //  total time of planes waiting to land
+	size_t takeoff_wait;             //  total time of planes waiting to take off
+	size_t idle_time;                //  total time runway is idle
 };
